backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.46
psycopg[binary]==3.3.2
python-dotenv==1.0.1
redis==7.1.1
pytest==8.4.2

